[CWS] move kfilters api transform functions to transform package
