+++ pkg/security/proto/api/transform/kfilters.go
@@ -3,22 +3,23 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-present Datadog, Inc.
 
-// Package api holds api related files
-package api
+// Package transform holds api types transformation related files
+package transform
 
 import (
 	"strconv"
 
 	"github.com/DataDog/datadog-agent/pkg/security/probe/kfilters"
+	"github.com/DataDog/datadog-agent/pkg/security/proto/api"
 	"github.com/DataDog/datadog-agent/pkg/security/secl/compiler/eval"
 	"github.com/DataDog/datadog-agent/pkg/security/secl/rules"
 )
 
 // FromProtoToFilterReport transforms a proto to a kfilter filter report
-func FromProtoToFilterReport(p *FilterReport) *kfilters.FilterReport {
+func FromProtoToFilterReport(p *api.FilterReport) *kfilters.FilterReport {
 	approverReports := make(map[eval.EventType]*kfilters.ApproverReport)
 
-	toAcceptModeRules := func(r []*AcceptModeRule) []kfilters.AcceptModeRule {
+	toAcceptModeRules := func(r []*api.AcceptModeRule) []kfilters.AcceptModeRule {
 		acceptModeRules := make([]kfilters.AcceptModeRule, len(r))
 		for i, rule := range r {
 			acceptModeRules[i] = kfilters.AcceptModeRule{
@@ -29,7 +30,7 @@ func FromProtoToFilterReport(p *FilterReport) *kfilters.FilterReport {
 	}
 
 	for _, report := range p.GetApprovers() {
-		approversToPrint := *report.GetApprovers().fromProtoToApprovers()
+		approversToPrint := fromProtoToApprovers(report.GetApprovers())
 		if len(approversToPrint) == 0 {
 			approversToPrint = nil // This is here to ensure that the printed result is `"Approvers": null` and not `"Approvers": {}`
 		}
@@ -46,7 +47,7 @@ func FromProtoToFilterReport(p *FilterReport) *kfilters.FilterReport {
 }
 
 // fromProtoToApprovers transforms a proto to a kfilter approvers
-func (p *Approvers) fromProtoToApprovers() *rules.Approvers {
+func fromProtoToApprovers(p *api.Approvers) rules.Approvers {
 	approvers := make(rules.Approvers)
 
 	for _, approver := range p.GetApproverDetails() {
@@ -68,15 +69,15 @@ func (p *Approvers) fromProtoToApprovers() *rules.Approvers {
 			})
 	}
 
-	return &approvers
+	return approvers
 }
 
 // FromFilterReportToProtoRuleSetReportMessage returns a pointer to a RuleSetReportMessage
-func FromFilterReportToProtoRuleSetReportMessage(filterReport *kfilters.FilterReport) *RuleSetReportMessage {
-	var reports []*ApproverReport
+func FromFilterReportToProtoRuleSetReportMessage(filterReport *kfilters.FilterReport) *api.RuleSetReportMessage {
+	var reports []*api.ApproverReport
 
 	for key, report := range filterReport.ApproverReports {
-		protoReport := &ApproverReport{
+		protoReport := &api.ApproverReport{
 			EventType:       key,
 			Mode:            uint32(report.Mode),
 			Approvers:       fromApproversToProto(report.Approvers),
@@ -86,18 +87,18 @@ func FromFilterReportToProtoRuleSetReportMessage(filterReport *kfilters.FilterRe
 		reports = append(reports, protoReport)
 	}
 
-	return &RuleSetReportMessage{
-		Filters: &FilterReport{
+	return &api.RuleSetReportMessage{
+		Filters: &api.FilterReport{
 			Approvers: reports,
 		},
 	}
 }
 
 // fromAcceptModeRulesToProto transforms a kfilter to a proto accept mode rules
-func fromAcceptModeRulesToProto(acceptModeRules []kfilters.AcceptModeRule) []*AcceptModeRule {
-	protoAcceptModeRules := make([]*AcceptModeRule, len(acceptModeRules))
+func fromAcceptModeRulesToProto(acceptModeRules []kfilters.AcceptModeRule) []*api.AcceptModeRule {
+	protoAcceptModeRules := make([]*api.AcceptModeRule, len(acceptModeRules))
 	for i, rule := range acceptModeRules {
-		protoAcceptModeRules[i] = &AcceptModeRule{
+		protoAcceptModeRules[i] = &api.AcceptModeRule{
 			RuleID: rule.RuleID,
 		}
 	}
@@ -105,8 +106,8 @@ func fromAcceptModeRulesToProto(acceptModeRules []kfilters.AcceptModeRule) []*Ac
 }
 
 // fromApproversToProto transforms a kfilter to a proto approvers
-func fromApproversToProto(approvers rules.Approvers) *Approvers {
-	protoApprovers := new(Approvers)
+func fromApproversToProto(approvers rules.Approvers) *api.Approvers {
+	protoApprovers := new(api.Approvers)
 
 	for field, filterValues := range approvers {
 		for _, filterValue := range filterValues {
@@ -114,13 +115,13 @@ func fromApproversToProto(approvers rules.Approvers) *Approvers {
 			stringFilterValue, ok := filterValue.Value.(string)
 			if !ok {
 				intFilterValue := filterValue.Value.(int)
-				protoApprovers.ApproverDetails = append(protoApprovers.ApproverDetails, &ApproverDetails{
+				protoApprovers.ApproverDetails = append(protoApprovers.ApproverDetails, &api.ApproverDetails{
 					Field: filterValue.Field,
 					Value: strconv.Itoa(intFilterValue),
 					Type:  int32(filterValue.Type),
 				})
 			} else {
-				protoApprovers.ApproverDetails = append(protoApprovers.ApproverDetails, &ApproverDetails{
+				protoApprovers.ApproverDetails = append(protoApprovers.ApproverDetails, &api.ApproverDetails{
 					Field: filterValue.Field,
 					Value: stringFilterValue,
 					Type:  int32(filterValue.Type),
